+++ packages/moshicam/src/constants.ts
@@ -1,0 +1,4 @@
+import { parseEther } from 'viem'
+
+export const MOSHIMINTER_ADMIN = '0xC701344E0e7fd31A7A348f0EA71ccb7F9833B107'
+export const DEFAULT_MINT_PRICE = parseEther('0.0001')